pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "onfra-pfas"
version = "0.1.0"
description = "ONFRA PFAS Non-Target Screening Desktop Application"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "김태형", email = "contact@onfra.org" }
]
keywords = ["PFAS", "mass-spectrometry", "NTS", "mzML", "OpenMS"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Chemistry",
]

dependencies = [
    "pyopenms>=3.0.0",
    "pyside6>=6.5.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyarrow>=14.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
gpu = [
    "cupy-cuda12x>=12.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pyinstaller>=6.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
onfra-pfas = "onfra_pfas.app.main:main"

[project.gui-scripts]
onfra-pfas-gui = "onfra_pfas.app.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/onfra_pfas"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "UP"]
